add rest of api endpoints
