--- api/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "target": "es2020",
-    "noEmit": true,
-    "lib": ["esnext"],
-    "types": ["node"]
-  },
-  "include": ["./**/*.ts"],
-  "exclude": []
-}
