+++ server/composables/supabase.ts
@@ -2,10 +2,16 @@ import { serverSupabaseServiceRole } from "#supabase/server"
 import { H3Event } from "h3"
 
 import { Database, Patch } from "~/lib/types"
+import type { UpdateSubscriptionInput } from "~/server/api/subscription.post"
 import { Logger } from "~/server/utils/logger"
 
 export const serverSupabase = (event: H3Event) => {
+  const config = useRuntimeConfig()
   const supabase = serverSupabaseServiceRole<Database>(event)
+
+  // -------------------------------------
+  // -------------- PATCHES --------------
+  // -------------------------------------
 
   const upsertPatches = async (patches: Patch[]) => {
     Logger.info(`Received ${patches.length} patches to maybe insert...`)
@@ -33,5 +39,88 @@ export const serverSupabase = (event: H3Event) => {
     }
   }
 
-  return { upsertPatches }
+  // -------------------------------------------
+  // -------------- SUBSCRIPTIONS --------------
+  // -------------------------------------------
+
+  const doesSubscriptionExist = async (endpoint: string): Promise<boolean> => {
+    const { data, error } = await supabase
+      .from("subscriptions")
+      .select("endpoint")
+      .eq("endpoint", endpoint)
+      .limit(1)
+      .maybeSingle()
+
+    if (error) {
+      throw new Error(error.message)
+    }
+
+    return data != null
+  }
+
+  const upsertSubscription = async ({
+    endpoint,
+    keys: { auth, p256dh },
+  }: UpdateSubscriptionInput) => {
+    const { error } = await supabase.from("subscriptions").upsert(
+      {
+        type: "push",
+        endpoint,
+        auth,
+        extra: p256dh,
+        environment: config.public.env,
+        lastNotified: 0,
+      },
+      { onConflict: "endpoint" },
+    )
+
+    if (error) {
+      throw new Error(error.message)
+    }
+  }
+
+  const deleteSubscription = async (endpoint: string) => {
+    const { error } = await supabase
+      .from("subscriptions")
+      .delete()
+      .eq("endpoint", endpoint)
+
+    if (error) {
+      throw new Error(error.message)
+    }
+  }
+
+  // ----------------------------------------------
+  // -------------- DISCORD WEBHOOKS --------------
+  // ----------------------------------------------
+
+  const registerDiscordWebhook = async (
+    endpoint: string,
+    guildId: string,
+    id: string,
+  ) => {
+    const { error } = await supabase.from("subscriptions").upsert(
+      {
+        type: "discord",
+        endpoint,
+        auth: guildId,
+        extra: id,
+        environment: config.public.env,
+        lastNotified: 0,
+      },
+      { onConflict: "endpoint" },
+    )
+
+    if (error) {
+      throw new Error(error.message)
+    }
+  }
+
+  return {
+    upsertPatches,
+    doesSubscriptionExist,
+    upsertSubscription,
+    deleteSubscription,
+    registerDiscordWebhook,
+  }
 }
